Api/requirements.txt
flask
flask-sqlalchemy 
mysql-connector-python
pandas